requirements.txt
urllib3~=1.24.3
requests~=2.20.0
beautifulsoup4~=4.9.3
PyYAML~=5.3.1
